% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRAN-and-GitHub-installation.R
\name{InstallRequiredPackages}
\alias{InstallRequiredPackages}
\title{Install required packages}
\usage{
InstallRequiredPackages(gitHubOnlyQ = FALSE)
}
\arguments{
\item{gitHubOnlyQ}{Should only the GitHub packages be installed or not?}
}
\description{
Installs required packages
}
\details{
The GitHub packages from "antononcube/R-packages" are
provide the recommender and LSA objects that showcased in the flexdashboard(s).
}
