% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProfileByFreeText.R
\name{ProfileByText}
\alias{ProfileByText}
\title{Profile by text}
\usage{
ProfileByText(
  smrObj,
  lsaObj,
  tagTypeWeights,
  text,
  nrecs = 10,
  representByTermsQ = TRUE,
  representByTopicsQ = TRUE
)
}
\arguments{
\item{smrObj}{An SMR object.}

\item{lsaObj}{An LSA object.}

\item{text}{Query text.}

\item{nrecs}{Number of the top recommended by \code{text} items
to be used for computing the profile.}

\item{representByTermsQ}{Should representation by terms be used or not?}

\item{representByTopicsQ}{Should representation by topics be used or not?}
}
\description{
Computes the profile of items in a SMR object that are retrieved
with a terms and topics profile derived using an LSA object.
}
\details{
Uses the \code{LSAMon} functions:
\code{\link{LSAMon::LSAMonRepresentByTerms}} and
\code{\link{LSAMon::LSAMonRepresentByTopics}}.
Uses the \code{SMRMon} functions:
\code{\link{SMRMon::SMRApplyTermWeightFunctions}},
\code{\link{SMRMon::SMRMonRecommendByProfile}}, and
\code{\link{SMRMon::SMRMonProfile}}
}
